내브바컴포넌트 다시분리
